Move HeartbeatController to a separate module
